First design of the OpenShift Pipelines next
I skipped the blah blah and went straight from the bones of the current state to
the meat of the technical implementation details.

I am interested on comments on goals and requirements, the implementation details
are just details, they are just here to syntethise feasibility in my head. It
is probably would be better to walk thru it over a VC call.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>